Refactor: change storybook folder
