+++ .storybook/preview-head.html
@@ -1,0 +1,1 @@
+<link href="https://fonts.googleapis.com/css?family=Muli:400,600,700|Rubik:400,500,700|Saira+Extra+Condensed:600" rel="stylesheet" />
