--- storybook/preview-head.html
@@ -1,7 +1,0 @@
-<link
-  rel="stylesheet"
-  href="https://unpkg.com/react-image-crop@4.0.1/dist/ReactCrop.css"
-/>
-<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-<link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,700" rel="stylesheet" />
-<link href="https://fonts.googleapis.com/css?family=Poppins:300,400,700" rel="stylesheet" />
