--- storybook/styles.css
@@ -1,22 +1,0 @@
-.buttonIconStyle {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: white;
-  font-size: 14px;
-}
-
-.iconStyle {
-  margin-right: 15px;
-}
-
-.customButton {
-  color: white;
-  font-family: 'Source Sans Pro', sans-serif;
-  font-size: 14px;
-  background-color: red;
-}
-
-.modalButton {
-  margin-right: 15px;
-}
